CCWebApp_api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.21
botocore==1.34.21
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
MarkupSafe==2.1.3
msgpack==1.0.7
openai==1.8.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
s3transfer==0.10.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
Werkzeug==3.0.1
zope.interface==6.1

django-webpack-loader==2.0.1
greenlet==3.0.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
packaging==23.2
priority==1.3.0
pyasn1==0.5.0
pyzmq==25.1.1
